requirements.txt
mss>=10.2.0
Pillow>=10.3.0
numpy>=1.23.5,<2.0.0
opencv-python-headless>=4.8.0
llama-cpp-python>=0.2.76
websockets>=13.1
onnxruntime>=1.16.3
pyclipper>=1.3.0.post6
